refractored bill page, into hook useBill, component BillTable, and for utility priceUtils
